chore: Moved models to self directories
