+++ personalization-sdk/src/main/kotlin/com/personalization/api/responses/products/filter/FilterValue.kt
@@ -1,4 +1,4 @@
-package com.personalization.api.responses.products
+package com.personalization.api.responses.products.filter
 
 import com.google.gson.annotations.SerializedName
