+++ personalization-sdk/src/main/kotlin/com/personalization/api/responses/products/image/ImageUrlResized.kt
@@ -1,4 +1,4 @@
-package com.personalization.api.responses.products
+package com.personalization.api.responses.products.image
 
 data class ImageUrlResized(
     val sizeToPath: Map<String, String>
