+++ personalization-sdk/src/main/kotlin/com/personalization/api/responses/products/price/PriceRangeItem.kt
@@ -1,4 +1,4 @@
-package com.personalization.api.responses.products
+package com.personalization.api.responses.products.price
 
 import com.google.gson.annotations.SerializedName
