+++ personalization-sdk/src/main/kotlin/com/personalization/api/responses/products/product/ProductParam.kt
@@ -1,4 +1,4 @@
-package com.personalization.api.responses.products
+package com.personalization.api.responses.products.product
 
 data class ProductParam(
     val key: String,
